+++ test.sh
@@ -21,6 +21,7 @@ swiftc -parse-as-library "$@" \
     Sources/AIUsage/CursorKeychain.swift \
     Sources/AIUsage/WorkSchedule.swift \
     Sources/AIUsage/Pace.swift \
+    Sources/AIUsage/ClaudeProfile.swift \
     Sources/AIUsage/Fetcher.swift \
     Sources/AIUsage/ProviderFolderAccess.swift \
     Sources/AIUsage/BrandGlyph.swift \